requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PyQtWebKit==5.15.2
requests==2.28.1
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.11
